Feature: hashtable - simplified system to not use different sized int

+++ include/int_hash_table.h
@@ -1,20 +1,20 @@
 #include <stdint.h>
 
 typedef struct hashItem {
-  uint16_t key;
-  uint32_t value;
+  int key;
+  int value;
 } HashItem;
 
 typedef struct hashTable {
-  uint16_t size;
-  uint16_t maxSize;
+  int size;
+  int maxSize;
   HashItem **hashItems;
 } HashTable;
 
 void freeHashTable(HashTable *hashTable);
 
-HashTable *createHashTable(uint16_t size);
+HashTable *createHashTable(int size);
 
-int8_t setHashItem(HashTable *hashTable, int16_t key, int32_t value);
+int setHashItem(HashTable *hashTable, int key, int value);
 
-HashItem *getHashItem(HashTable *hashTable, int16_t key);
+HashItem *getHashItem(HashTable *hashTable, int key);
